feat: move resources-v1 at the same level that resources-v2
