+++ gravitee-apim-console-webui/src/management/api/_api.scss
@@ -4,8 +4,8 @@
 @import 'audit/history/apiHistory';
 @import 'policy-studio-v1/policies/policy';
 @import 'proxy/properties-v1/properties';
-@import 'proxy/resources-v1/resources';
 @import 'proxy/health-check-dashboard/healthcheck';
+@import 'resources-v1/resources';
 
 .gravitee-api-container {
   border: 1px solid;
