+++ gravitee-apim-console-webui/src/management/api/proxy/api-proxy.module.ts
@@ -23,10 +23,10 @@ import { ApiProxyFailoverModule } from './failover/api-proxy-failover.module';
 import { ApiProxyHealthCheckModule } from './health-check/api-proxy-health-check.module';
 import { ApiProxyHealthCheckDashboardModule } from './health-check-dashboard/api-proxy-health-check-dashboard.module';
 import { ApiV1PropertiesComponent } from './properties-v1/properties.component';
-import { ApiV1ResourcesComponent } from './resources-v1/resources.component';
 import { ApiPropertiesModule } from './properties/properties/api-properties.module';
 import { ApiDynamicPropertiesModule } from './properties/dynamic-properties/api-dynamic-properties.module';
 
+import { ApiV1ResourcesComponent } from '../resources-v1/resources.component';
 import { ApiProxyCorsModule } from '../cors/api-proxy-cors.module';
 
 @NgModule({
