+++ gravitee-apim-console-webui/src/management/api/resources-v1/_resources.scss
