+++ gravitee-apim-console-webui/src/management/api/resources-v1/resources.component.ajs.ts
