+++ gravitee-apim-console-webui/src/management/api/resources-v1/resources.controller.ajs.ts
@@ -132,7 +132,7 @@ class ApiV1ResourcesControllerAjs {
       .show({
         controller: 'DialogConfirmController',
         controllerAs: 'ctrl',
-        template: require('html-loader!../../../../components/dialog/confirmWarning.dialog.html'),
+        template: require('html-loader!../../../components/dialog/confirmWarning.dialog.html'),
         clickOutsideToClose: true,
         locals: {
           title: 'Are you sure you want to remove this resource?',
